+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## NEXT
 * improved error handling.
+* new method GetHealth().
 
 ## 1.1.0 (2020-06-25)
 * supports FOSSology REST API v1.0.16 (FOSSology 3.8.0 built @ 2020/06/19).